requirements.txt
openai
python-dotenv
azure-cognitiveservices-speech
# ...other dependencies if needed...